Split repository.ts: extract the dev-team watches domain (audit L14, 2/N)
Second slice of the repository.ts carve-up started in #804. Moves the
dev-team completion-DM watch queries — DevTeamWatchInput, DevTeamWatch,
insertDevTeamWatch, listUnnotifiedDevTeamWatches, markDevTeamWatchNotified —
into src/storage/repository/devTeamWatches.ts.

Pure code motion: all five symbols are byte-identical to their previous
definitions (verified mechanically against origin/main), and repository.ts
re-exports the new module alongside the existing two, so every import site
and tests/repository.test.ts are untouched.

Chosen as the second domain for the same reason as the first two: the cluster
is fully self-contained (its only dependencies are `pool` and the `Platform`
type), it owns one table nobody else touches, and it carries no
conversation-scoped admin read — so the 🔒 SQL-scoping invariant is not in
play and the diff stays reviewable as a move.

Per the lesson #804 learned the hard way, I checked tests/ for source-text
introspection before removing the section banner: the only test that slices
repository.ts is the member-projects HTTP scan, which #804 already made
banner-agnostic, and the dev-team banner sat at the end of the file.
